+++ network-area-diagram/src/test/java/com/powsybl/nad/svg/NominalVoltageStyleTest.java
@@ -99,27 +99,6 @@ public class NominalVoltageStyleTest extends AbstractTest {
     }
 
     @Test
-    public void testEurope() {
-        Network network = Network.read("simple-eu.uct", getClass().getResourceAsStream("/simple-eu.uct"));
-        LoadFlow.run(network);
-        assertEquals(toString("/simple-eu.svg"), generateSvgString(network, "/simple-eu.svg"));
-    }
-
-    @Test
-    public void testEuropeLoopAperture80() {
-        Network network = Network.read("simple-eu.uct", getClass().getResourceAsStream("/simple-eu.uct"));
-        getSvgParameters().setLoopEdgesAperture(80);
-        assertEquals(toString("/simple-eu-loop80.svg"), generateSvgString(network, "/simple-eu-loop80.svg"));
-    }
-
-    @Test
-    public void testEuropeLoopAperture100() {
-        Network network = Network.read("simple-eu.uct", getClass().getResourceAsStream("/simple-eu.uct"));
-        getSvgParameters().setLoopEdgesAperture(100);
-        assertEquals(toString("/simple-eu-loop100.svg"), generateSvgString(network, "/simple-eu-loop100.svg"));
-    }
-
-    @Test
     public void testHvdc() {
         Network network = FourSubstationsNodeBreakerFactory.create();
         assertEquals(toString("/hvdc.svg"), generateSvgString(network, "/hvdc.svg"));
